[Backend] Improved resource descriptors and hooks - Fixed invalid SetDescriptorCommand command type - Variable length SetDescriptorCommand - Added create/destroy resource hooks
[DX12, Vulkan]
- Added subresource and region indexing to all descriptor regions

+++ Source/Libraries/Backend/Include/Backend/Resource/ResourceToken.h
@@ -35,7 +35,7 @@
 /// Unpacked token type
 struct ResourceToken {
     /// Physical UID of the resource
-    uint64_t puid;
+    uint32_t puid;
 
     /// Type of the resource
     Backend::IL::ResourceTokenType type;
